Moved get_dd4 in the RiskComputer
