+++ openquake/risklib/tests/scientific_test.py
@@ -18,11 +18,12 @@
 
 import unittest
 import pickle
+import toml
 
 import numpy
 import pandas
 import matplotlib.pyplot as plt
-from openquake.risklib import scientific
+from openquake.risklib import riskmodels, scientific
 
 PLOTTING = False
 aae = numpy.testing.assert_allclose
@@ -642,3 +643,102 @@ class ComposeDDSTestCase(unittest.TestCase):
         dd = comp([comp([dd1, dd2]), dd3])
         aae(exp, dd)
         aae(exp, comp([dd1, dd2, dd3]))
+
+
+class RiskComputerTestCase(unittest.TestCase):
+    def test1(self):
+        dic = {'asset_df': {'area': [1.0],
+                            'id': [b'a2'],
+                            'lat': [38.17],
+                            'lon': [15.56],
+                            'site_id': [0],
+                            'taxonomy': [1],
+                            'value-number': [2000.0],
+                            'value-structural': [2000.0]},
+               'calculation_mode': 'event_based_risk',
+               'loss_types': ['structural'],
+               'minimum_asset_loss': {'structural': 0},
+               'risk_functions': {
+                   'RC#structural':
+                   {"openquake.risklib.scientific.VulnerabilityFunction":
+                    {"id": "RC",
+                     "peril": 'earthquake',
+                     "loss_type": "structural",
+                     "imt": "PGA",
+                     "imls": [0.1, 0.2, 0.3, 0.5, 0.7],
+                     "mean_loss_ratios": [0.0035, 0.07, 0.14, 0.28, 0.56],
+                     "covs": [0.0, 0.0, 0.0, 0.0, 0.0],
+                     "distribution_name": "LN"}}},
+               'wdic': {'RC#structural': 1}}
+        gmfs = {'eid': [0, 1],
+                'sid': [0, 0],
+                'gmv_0': [.23, .31]}
+        rc = riskmodels.get_riskcomputer(dic, alias={'PGA': 'gmv_0'})
+        print(toml.dumps(dic))
+        for k, v in rc.todict().items():
+            self.assertEqual(dic[k], v)
+        out = rc.output(pandas.DataFrame(gmfs))
+        print(out)
+
+    def test2(self):
+        alias = {'PGA': 'gmv_0',
+                 'SA(0.2)': 'gmv_1',
+                 'SA(0.5)': 'gmv_2',
+                 'SA(0.8)': 'gmv_3',
+                 'SA(1.0)': 'gmv_4'}
+        dic = {'asset_df': {'area': [10.0, 1.0],
+                            'id': [b'a0', b'a3'],
+                            'lat': [29.1098, 27.9015],
+                            'lon': [81.2985, 85.7477],
+                            'policy': [2, 1],
+                            'site_id': [0, 2],
+                            'taxonomy': [1, 1],
+                            'value-nonstructural': [1500.0, 2500.0],
+                            'value-number': [3.0, 10.0],
+                            'value-structural': [3000.0, 5000.0]},
+               'calculation_mode': 'event_based_risk',
+               'loss_types': ['nonstructural', 'structural'],
+               'risk_functions': {
+                   'RM#nonstructural': {
+                       'openquake.risklib.scientific.VulnerabilityFunction': {
+                           'covs': [0.0001, 0.0001, 0.0001, 0.0001, 0.0001],
+                           'distribution_name': 'LN',
+                           'id': 'RM',
+                           'imls': [0.1, 0.2, 0.4, 0.7, 1.0],
+                           'imt': 'SA(1.0)',
+                           'mean_loss_ratios': [0.1, 0.2, 0.35, 0.6, 0.9]}},
+                   'RM#structural': {
+                       'openquake.risklib.scientific.VulnerabilityFunction': {
+                           'covs': [0.0001, 0.0001, 0.0001, 0.0001, 0.0001],
+                           'distribution_name': 'LN',
+                           'id': 'RM',
+                           'imls': [0.02, 0.3, 0.5, 0.9, 1.2],
+                           'imt': 'PGA',
+                           'mean_loss_ratios': [0.05, 0.1, 0.2, 0.4, 0.8]}}},
+               'wdic': {'RM#nonstructural': 1, 'RM#structural': 1}}
+
+        gmfs = {'eid': [0, 2],
+                'sid': [0, 0],
+                'gmv_0': [.23, .31],
+                'gmv_1': [.23, .41],
+                'gmv_2': [.23, .51],
+                'gmv_3': [.23, .32],
+                'gmv_4': [.23, .21]}
+        rc = riskmodels.get_riskcomputer(dic, alias)
+        print(toml.dumps(dic))
+        out = rc.output(pandas.DataFrame(gmfs))
+        print(out)
+
+    def test3(self):
+        # TODO: a case with nontrivial wdic, for instance
+        # 'wdic': {'Wood/Com#contents': 1.0,
+        #          'Wood1#structural': 0.6,
+        #          'Wood2#structural': 0.4}
+        # NB: for the same loss type the weights must sum up to 1
+        pass
+
+    def test4(self):
+        # TODO: a case with secondary losses, for instance
+        # seclosses = [partial(total_losses, kind=oq.total_losses)]
+        # seclosses = [partial(insurance_losses, policy_df=self.policy_df)]
+        pass
